// src/app/shared/models/Package/single-package-print-scan-model.ts
export class SinglePackagePrintScanModel {
  packageNumber: string;
  orderNumbers: string[];
  skuList: string[];
  pWeight: number;
  isExcept: boolean;
  skuImg: string;
  quantity: number;
  packageStatusText: string;
  status: string;
  type: string;
  packageTypeDesc: string;
  totalQuantity: number;
  trackNumber: string;
  packageId: number;
}